@@ -1,13 +1,23 @@
-import React, { createContext, useContext } from 'react';
+import React, {
+  createContext, useContext, useState, useEffect,
+} from 'react';
 import PropTypes from 'prop-types';
+import { useFetchUser } from '@/lib/user';
 
 export const AppContext = createContext();
 
 const AppContextProvider = ({ children }) => {
-  const title = 'Test';
+  const { loading, user } = useFetchUser();
+  const [currentUser, setCurrentUser] = useState();
+
+  useEffect(() => {
+    if (user && !loading) {
+      setCurrentUser(user);
+    }
+  }, [user]);
 
   return (
-    <AppContext.Provider value={{ title }}>
+    <AppContext.Provider value={{ currentUser, setCurrentUser }}>
       {children}
     </AppContext.Provider>
   );
